requirements.txt
pygame-ce
numpy
pygame-gui
cupy-cuda11x
